FIXES #31 完成RPC客户端向多个服务器写入的操作，支持异步模式调用

+++ monad-interface/src/main/scala/monad/face/services/RpcFacade.scala
@@ -31,7 +31,7 @@ trait RpcSearcherFacade {
   def collectSearch2(resourceName: String, q: String, sort: String, topN: Int): ShardResult
 
   @Rpc(mode = "all", merge = "collectMaxDoc")
-  def maxDoc(resourceName: String): Int
+  def maxDoc(resourceName: String): Long
 
   /**
    * 搜索对象
